requirements.txt
opencv-python==4.8.1.78
requests==2.31.0
azure-cosmos==4.5.1
numpy==1.24.3
ultralytics==8.0.196
torch==2.1.0
torchvision==0.16.0
